refactor apropiation code for handle errors and add some helper function's over that poroccess

+++ models/rubro.go
@@ -6,7 +6,7 @@ type Rubro struct {
 	Entidad         int              `orm:"column(entidad)"`
 	Codigo          string           `orm:"column(codigo)"`
 	Descripcion     string           `orm:"column(descripcion);null"`
-	UnidadEjecutora int16            `orm:"column(unidad_ejecutora)"`
+	UnidadEjecutora int              `orm:"column(unidad_ejecutora)"`
 	Nombre          string           `orm:"column(nombre);null"`
 	ProductoRubro   []*ProductoRubro `orm:"reverse(many)"`
 }
